Refactor: separate the creation of informers from tunnel server component
